functionality to skip an op in the oplog
Summary:
A new configuration option replSetSkipOpId is added
to mongod which makes secondarys skip an op received from
the primary's oplog whose uid (which is the "h" field in the
oplog collection) is equal to replSetSkipOpId.

Note on a secondary's crash, mongod logs the op information
in the error log. That op information can be used to pass
a value for --replSetSkipOpId.

Test Plan:
jstest to verify secondary skips an op inserted in
primary's oplog when run with the option --replSetSkipOpId.
Verified the output log of the test to check the log message
added in this diff.